requirements.txt
aniso8601==7.0.0
arrow==0.13.0
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.5
beautifulsoup4==4.7.1
binary==1.0.0
blessed==1.15.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
cryptography==2.4.2
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.10
django-allauth==0.38.0
django-braces==1.13.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-currentuser==0.2.3
django-extensions==2.1.4
django-filter==2.1.0
django-graphql-jwt==0.3.0
django-lint==2.0.4
django-model-utils==3.1.2
django-nose==1.4.6
django-notifications-hq==1.5.0
django-oauth-toolkit==1.2.0
django-picklefield==2.0
django-q==1.0.1
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-social-share==1.3.0
djangorestframework==3.9.0
docopt==0.6.2
dpath==1.4.2
drf-yasg==1.12.1
env==0.1.0
facebook-sdk==3.1.0
flake8==3.6.0
future==0.17.1
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
graphene==2.1.8
graphene-django==2.8.2
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
json-spec==0.10.1
jsonfield==2.0.2
jsonschema==2.6.0
lazy-object-proxy==1.3.1
logilab-astng==0.24.3
logilab-common==1.4.2
lxml==4.3.0
markdown2==2.3.7
MarkupSafe==1.1.0
mccabe==0.6.1
nose==1.3.7
notifications==0.3.2
oauthlib==3.0.0
openapi-codec==1.3.2
pep8==1.7.1
Pillow==5.4.1
pip-tools==3.2.0
promise==2.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyquery==1.4.0
python-dateutil==2.7.5
python-decouple==3.1
python-twitter==3.5
python3-openid==3.1.0
pytz==2018.9
PyYAML==4.2b1
readtime==1.0.6
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
ruamel.yaml==0.15.86
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
soupsieve==1.7.3
swagger-py-codegen==0.3.9
swagger-spec-validator==2.4.3
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.6.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.0
